[](refact) update references to common interfaces
