+++ application/apps/indexer-neon/package.json
@@ -2,8 +2,8 @@
   "name": "chipmunk.indexer.neon",
   "version": "0.0.1",
   "description": "",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./dist/apps/indexer-neon/src/index.js",
+  "types": "./dist/apps/indexer-neon/src/index.d.ts",
   "scripts": {
     "neon": "npm run build-ts-neon && npm run build-neon",
     "build-ts-neon": "tsc -p ./tsconfig.json",
