+++ application/apps/indexer-neon/src/merger.ts
@@ -1,4 +1,3 @@
-const addon = require("../native");
 import { log } from "./logging";
 import {
     ITicks,
