+++ application/apps/indexer-neon/src/native.ts
@@ -1,0 +1,5 @@
+
+export function getNativeModule() {
+    const native = require("../../../../native/index.node");
+    return native;
+}
